requirements.txt
asgiref==3.8.1
colorama==0.4.6
Django==5.1.5
django-cors-headers==4.6.0
django-dotenv==1.4.2
django-phonenumber-field==8.0.0
djangorestframework==3.15.2
djangorestframework_simplejwt==5.4.0
mysqlclient==2.2.7
pillow==11.1.0
PyJWT==2.10.1
python-dotenv==1.0.1
qrcode==8.0
sqlparse==0.5.3
tzdata==2025.1
